refactor: move type to file
