+++ src/types/auth/index.ts
@@ -1,3 +1,1 @@
-export * from '../user/response/userInfoResponse';
-export * from './response/authResponse';
-export * from './response/kakaoUserInfoResponse';
+export * from './response';
