+++ src/types/auth/response.ts
@@ -1,3 +1,10 @@
+export type AuthResponse = {
+  accessToken: string;
+  refreshToken: string;
+  userId: number;
+  accessTokenExpireDate: number;
+};
+
 export type KakaoProperties = {
   nickname: string;
   profile_image: string;
