--- src/types/auth/response/authResponse.ts
@@ -1,6 +1,0 @@
-export type AuthResponse = {
-  accessToken: string;
-  refreshToken: string;
-  userId: number;
-  accessTokenExpireDate: number;
-};
