--- src/types/community/communityIdCardSummary.ts
@@ -1,3 +1,0 @@
-import { IdCardDetail } from '@/types/idCard';
-
-export type CommunityIdCardSummary = Omit<IdCardDetail, 'profileImageUrl'>;
