+++ src/types/community/dto.ts
@@ -1,3 +1,7 @@
+import { IdCardDetail } from '@/types/idCard';
+
+export type CommunityIdCardSummary = Omit<IdCardDetail, 'profileImageUrl'>;
+
 export type CommunitySummary = {
   communityId: number;
   thumbnailImageUrl: string;
