+++ src/types/community/index.ts
@@ -1,4 +1,2 @@
-export * from './communityIdCardSummary';
-export * from './communitySummary';
-export * from './response/communitiesResponse';
-export * from './response/communityIdCardsResponse';
+export * from './dto';
+export * from './response';
