+++ src/types/community/response.ts
@@ -1,0 +1,7 @@
+import { SliceResponse } from '@/types/api';
+import { CommunitySummary } from '@/types/community';
+import { CommunityIdCardSummary } from '@/types/community';
+
+export type CommunitiesResponse = CommunitySummary[];
+
+export type CommunityIdCardsResponse = SliceResponse<CommunityIdCardSummary>;
