--- src/types/community/response/communitiesResponse.ts
@@ -1,3 +1,0 @@
-import { CommunitySummary } from '@/types/community';
-
-export type CommunitiesResponse = CommunitySummary[];
