--- src/types/community/response/communityIdCardsResponse.ts
@@ -1,4 +1,0 @@
-import { SliceResponse } from '@/types/api';
-import { CommunityIdCardSummary } from '@/types/community/communityIdCardSummary';
-
-export type CommunityIdCardsResponse = SliceResponse<CommunityIdCardSummary>;
