--- src/types/idCard/createKeyword.ts
@@ -1,3 +1,0 @@
-import { Keyword } from '@/types/idCard';
-
-export type CreateKeyword = Omit<Keyword, 'keywordId'>;
