+++ src/types/idCard/index.ts
@@ -1,4 +1,3 @@
-export * from './createKeyword';
-export * from './idCardDetail';
-export * from './keyword';
-export * from './response/commentCountResponse';
+export * from './dto';
+export * from './request';
+export * from './response';
