--- src/types/idCard/keyword.ts
@@ -1,6 +1,0 @@
-export type Keyword = {
-  keywordId: number;
-  title: string;
-  imageUrl: string;
-  content: string;
-};
