+++ src/types/idCard/response.ts
@@ -1,3 +1,7 @@
 import { IdCardDetail } from '@/types/idCard';
 
+export type CommentCountResponse = {
+  count: number;
+};
+
 export type IdCardDetailResponse = IdCardDetail;
