--- src/types/idCard/response/commentCountResponse.ts
@@ -1,3 +1,0 @@
-export type CommentCountResponse = {
-  count: number;
-};
