+++ src/types/user/index.ts
@@ -1,1 +1,3 @@
-export * from './response/userInfoResponse';
+export * from './dto';
+export * from './request';
+export * from './response';
