+++ src/types/user/request.ts
@@ -1,4 +1,4 @@
 // user.d.ts 정리 후 절대경로 적용
-import { UserInfoResponse } from '../response/userInfoResponse';
+import { UserInfoResponse } from './response';
 
 export type UserInfoRequest = Omit<UserInfoResponse, 'userId'>;
