+++ src/types/user/response.ts
@@ -1,0 +1,3 @@
+import { UserInfo } from './dto';
+
+export type UserInfoResponse = UserInfo;
